+++ api/serializers/__init__.py
@@ -5,3 +5,4 @@ from .new_threshold_field import NewThresholdField
 from .app_bookmark_field import AppBookmarkField
 from .tag_related_field import TagRelatedField
 from .identity_related_field import IdentityRelatedField
+from .instance_related_field import InstanceRelatedField
